requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==25.0
Pillow==10.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
